+++ src/main/java/org/vanilladb/bench/benchmarks/micro/rte/jdbc/CheckDatabaseJdbcJob.java
@@ -25,7 +25,7 @@ import java.util.logging.Logger;
 import org.vanilladb.bench.remote.SutResultSet;
 import org.vanilladb.bench.remote.jdbc.VanillaDbJdbcResultSet;
 import org.vanilladb.bench.rte.jdbc.JdbcJob;
-import org.vanilladb.bench.server.param.micro.MicroTestbedLoaderSpHelper;
+import org.vanilladb.bench.server.procedure.micro.MicroTestbedLoaderSpHelper;
 
 public class CheckDatabaseJdbcJob implements JdbcJob {
 	private static Logger logger = Logger.getLogger(CheckDatabaseJdbcJob.class.getName());
