+++ src/main/java/org/vanilladb/bench/server/procedure/micro/MicroCheckDatabaseProc.java
@@ -3,7 +3,6 @@ package org.vanilladb.bench.server.procedure.micro;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import org.vanilladb.bench.server.param.micro.MicroTestbedLoaderSpHelper;
 import org.vanilladb.bench.server.procedure.StoredProcedureUtils;
 import org.vanilladb.core.query.algebra.Scan;
 import org.vanilladb.core.sql.storedprocedure.StoredProcedure;
