+++ src/main/java/org/vanilladb/bench/server/procedure/micro/MicroTxnSpHelper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package org.vanilladb.bench.server.param.micro;
+package org.vanilladb.bench.server.procedure.micro;
 
 import org.vanilladb.bench.benchmarks.micro.rte.MicroTxnParamHelper;
 import org.vanilladb.core.sql.DoubleConstant;
