+++ src/main/java/org/vanilladb/bench/server/procedure/tpce/TradeOrderSpHelper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package org.vanilladb.bench.server.param.tpce;
+package org.vanilladb.bench.server.procedure.tpce;
 
 import org.vanilladb.bench.benchmarks.tpce.rte.TradeOrderParamHelper;
 import org.vanilladb.core.sql.BigIntConstant;
