+++ src/main/java/org/vanilladb/bench/server/procedure/ycsb/YcsbTxSpHelper.java
@@ -1,4 +1,4 @@
-package org.vanilladb.bench.server.param.ycsb;
+package org.vanilladb.bench.server.procedure.ycsb;
 
 import org.vanilladb.bench.benchmarks.ycsb.YcsbConstants;
 import org.vanilladb.bench.benchmarks.ycsb.rte.YcsbTxParamHelper;
